Add new GeoServiceManager module
#318
