+++ modules/layer/__init__.py
@@ -1,2 +1,3 @@
 from .add_layer import LayerAdder
+from .geo_service import GeoServiceManager
 from .limitations_checker import LimitationsChecker
